backend/requirements.txt
# Кроп + превью
python-slugify==8.0.1
gunicorn==21.2.0

Django==5.2.1
django-filter==25.1
django-cors-headers==4.7.0
pillow==11.2.1
asgiref==3.8.1
psycopg2-binary>=2.9.6,<3.0.0
django-nested-admin>=4.0.2
requests==2.32.4
